Fix warning around type coercion on 64 bit windows; range is explicitly checked

+++ Modules/_io/fileio.c
@@ -751,7 +751,7 @@ _io_FileIO_readall_impl(fileio *self)
             Py_END_ALLOW_THREADS
 
             if (end >= pos && pos >= 0 && end - pos < PY_SSIZE_T_MAX) {
-                bufsize = bufsize - pos;
+                bufsize = bufsize - (size_t)(pos);
             }
         }
     }
